# Conflicts:
#	lib/models/countries_model.dart
#	lib/pages/main_page.dart
#	lib/repositories/country_repos.dart
#	lib/services/storage_service.dart

+++ lib/services/storage_service.dart
@@ -7,21 +7,14 @@ import 'package:sqflite/sqflite.dart';
 import '../models/countries_model.dart';
 
 class StorageService {
+  static List<Map> sqlResponse = [];
   Future<bool> getCountriesFromMemory() async {
     if(await databaseExists(CacheKeys.databasePath)) {
       try{
-        Database database = await openDatabase(
-            CacheKeys.databasePath,
-            version: 1,
-            onCreate: (Database db, int version) async {
-              // When creating the db, create the table
-              await db.execute(
-                  'CREATE TABLE countries (country varchar(255))');
-            }
-        );
-        List<Map> sqlResponse = await database.rawQuery('SELECT * FROM countries');
-        CachedModels.countries = countryModelFromJson({"data": {"country": sqlResponse.toString()}}.toString());
-        return true;
+        Database database = await openDatabase(CacheKeys.databasePath);
+        sqlResponse = await database.rawQuery("SELECT * FROM countries;");
+        CachedModels.countries = countryModelFromJson(sqlResponse);
+        return sqlResponse.isNotEmpty;
       } catch(error, stacktrace) {
         debugPrint("$error, $stacktrace");
         return false;
@@ -31,12 +24,25 @@ class StorageService {
     }
   }
 
-  Future<bool> writeCountriesToMemory(CountryModel countries) async {
+  Future<bool> writeCountriesToMemory(List<CountryModel> countries) async {
    try{
      Database database = await openDatabase(CacheKeys.databasePath);
-     database.delete('countries');
-     for(Country item in countries.data.countries) {
-       database.insert('countries', {"country": countryToJson(item)});
+     List<Map<String, dynamic>> countriesJson = countryModelToJson(countries);
+     if((await database.rawQuery("SELECT * FROM sqlite_master WHERE name = 'countries' and type='table'")).isEmpty) {
+       await database.execute("CREATE TABLE countries ("
+           "code TEXT,"
+           "name TEXT,"
+           "phone TEXT,"
+           "continentCode TEXT,"
+           "continentName TEXT,"
+           "capital TEXT,"
+           "currency TEXT,"
+           "emoji TEXT)");
+     } else {
+       database.delete('countries');
+     }
+     for(Map<String, dynamic> item in countriesJson) {
+       database.insert('countries', item);
      }
      return true;
    } catch (error, stacktrace) {
